+++ plugins/codesnippet/lang/zh.js
@@ -8,5 +8,6 @@ CKEDITOR.plugins.setLang( 'codesnippet', 'zh', {
 	codeContents: '程式碼內容',
 	emptySnippetError: '程式碼片段不可為空白。',
 	language: '語言',
-	title: '程式碼片段'
+	title: '程式碼片段',
+	pathName: 'codesnippet' // MISSING
 } );